fix(dm-review): resolve run-cost-summary review findings (P2+P3)
Strengthen validate_run_cost_summary to enforce exact key sets,
usage_provenance structure, integer coverage types, digest type, and
volatile_fields equality (rejects forged exclusions). Move run-cost-summary
code to its own cost_summary.py module; build_run_cost_summary now leaves
digest None (CLI is sole finalizer after redaction). Split the builder
into focused helpers. Add dirty-state auto-detection to all 11 prescribed
Markdown wiring sites. Replace tautological digest test with real CLI
end-to-end verification; add --repository-commit/--dirty-state CLI tests
and 7 validator rejection tests. Clarify volatile_fields docs.

29/29 run-cost-summary tests, 988/988 suite, validator 13/13,
composition/dual-compat/codex-shims all PASS.
