+++ plugins/pipeline/skills/pipeline-run/SKILL.md
@@ -217,10 +217,10 @@ After the authoritative terminal receipt is appended to the cumulative receipt a
 "$WORKFLOW_KERNEL" observe-pipeline --manifest plans/<feature>/manifest.json --receipts plans/<feature>/authoritative-receipts.json --state-dir plans/<feature>
 "$WORKFLOW_KERNEL" compare --state-dir plans/<feature> --authoritative-receipts plans/<feature>/authoritative-receipts.json --output plans/<feature>/shadow-report.json
 "$WORKFLOW_KERNEL" metrics --events plans/<feature>/authoritative-receipts.json --output plans/<feature>/metrics.json
-"$WORKFLOW_KERNEL" run-cost-summary --events plans/<feature>/authoritative-receipts.json --output plans/<feature>/run-cost-summary.json --repository-commit "$(git rev-parse HEAD)"
+"$WORKFLOW_KERNEL" run-cost-summary --events plans/<feature>/authoritative-receipts.json --output plans/<feature>/run-cost-summary.json --repository-commit "$(git rev-parse HEAD)" $(test -n "$(git status --porcelain)" && echo --dirty-state)
 ```
 
-The `run-cost-summary` command emits a schema-bound `run-cost-summary.json` artifact beside the authoritative receipts. It is observation-only and never gates, waives, or alters any lane, phase, or review outcome. If the kernel runtime is unavailable, the command is a silent no-op — record `run-cost-summary unavailable` as a skip reason in the progress ledger and continue delivery; never block on it. Pass `--dirty-state` when the working tree has uncommitted changes.
+The `run-cost-summary` command emits a schema-bound `run-cost-summary.json` artifact beside the authoritative receipts. It is observation-only and never gates, waives, or alters any lane, phase, or review outcome. If the kernel runtime is unavailable, the command is a silent no-op — record `run-cost-summary unavailable` as a skip reason in the progress ledger and continue delivery; never block on it. The command auto-detects a dirty working tree via `git status --porcelain` and passes `--dirty-state` accordingly.
 
 `bind-prediction` runs before corresponding authoritative actions, atomically seals the source, translated events, event digest, and RunSpec context, and appends exact binding authority to the canonical lifecycle ledger before `run.started`. `observe-pipeline` runs only after authoritative receipts exist and requires that ordered authority plus bound `pipeline-shadow-prediction.json`; direct comparison rechecks the same authority, and byte-identical predicted and authoritative receipts are valid only with the pre-start binding. It writes a separate `authoritative_observation` and never creates or changes the prediction. Without matching independent prediction evidence, observation and comparison fail closed. Keep the source and bound artifact until comparison completes. Write the shadow report and reliability metrics without changing the merge recommendation, cleanup disposition, or provider result. Never auto-delete the repository-lifetime `.workflow-kernel/repository-scope.json`; retain the terminal run directory or a durable tombstone until fresh exact-scope Docker inventory proves zero exact-run objects and no uninspectable matches, regardless of parity `match`.
