+++ plugins/pipeline/skills/pipeline/SKILL.md
@@ -157,10 +157,10 @@ After the authoritative terminal receipt is appended, invoke exactly:
 "$WORKFLOW_KERNEL" observe-pipeline --manifest plans/<feature-slug>/manifest.json --receipts plans/<feature-slug>/authoritative-receipts.json --state-dir plans/<feature-slug>
 "$WORKFLOW_KERNEL" compare --state-dir plans/<feature-slug> --authoritative-receipts plans/<feature-slug>/authoritative-receipts.json --output plans/<feature-slug>/shadow-report.json
 "$WORKFLOW_KERNEL" metrics --events plans/<feature-slug>/authoritative-receipts.json --output plans/<feature-slug>/metrics.json
-"$WORKFLOW_KERNEL" run-cost-summary --events plans/<feature-slug>/authoritative-receipts.json --output plans/<feature-slug>/run-cost-summary.json --repository-commit "$(git rev-parse HEAD)"
+"$WORKFLOW_KERNEL" run-cost-summary --events plans/<feature-slug>/authoritative-receipts.json --output plans/<feature-slug>/run-cost-summary.json --repository-commit "$(git rev-parse HEAD)" $(test -n "$(git status --porcelain)" && echo --dirty-state)
 ```
 
-The `run-cost-summary` command emits a schema-bound `run-cost-summary.json` artifact beside the authoritative receipts. It is observation-only and never gates, waives, or alters any lane, phase, or review outcome. If the kernel runtime is unavailable, record `run-cost-summary unavailable` as a skip reason in the progress ledger and continue; never block delivery. Pass `--dirty-state` when the working tree has uncommitted changes.
+The `run-cost-summary` command emits a schema-bound `run-cost-summary.json` artifact beside the authoritative receipts. It is observation-only and never gates, waives, or alters any lane, phase, or review outcome. If the kernel runtime is unavailable, record `run-cost-summary unavailable` as a skip reason in the progress ledger and continue; never block delivery. The command auto-detects a dirty working tree via `git status --porcelain` and passes `--dirty-state` accordingly.
 
 `bind-prediction` atomically seals the prediction source, translated events, event digest, and RunSpec context as `pipeline-shadow-prediction.json` and appends exact binding authority to the canonical lifecycle ledger before `run.started`. Observation and direct comparison both require that ordered binding and matching artifact; byte-identical predicted and authoritative receipts are valid only with this durable pre-start proof. `observe-pipeline` writes a separate `authoritative_observation` and never creates or changes the prediction. Keep both the source and bound artifact through comparison. Missing or mismatched independent prediction evidence fails closed. Exit `5` is a visible parity gap, not a pipeline failure; observation/runtime failures preserve the canonical result. `.workflow-kernel/repository-scope.json` is repository-lifetime durable and never auto-deleted. Parity `match` alone never deletes the terminal run directory; retain it or a durable tombstone until fresh exact-scope Docker inventory proves zero exact-run objects and no uninspectable matches.
